dbpush and get

+++ src/main/java/com/project/repository/ProductRepository.java
@@ -2,6 +2,7 @@ package com.project.repository;
 
 import com.project.entity.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
+import java.util.Optional;
 
-public interface ProductRepository extends JpaRepository<Product, Integer> {
+public interface ProductRepository extends JpaRepository<Product, Long> {
 }
